+++ src/main/java/pe/edu/upc/greenly/repositories/CampañaFavoritaRepository.java
@@ -3,5 +3,5 @@ package pe.edu.upc.greenly.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pe.edu.upc.greenly.entities.CampañaFavorita;
 
-public interface CampañaFavoritaRepository extends JpaRepository<CampañaFavorita, Integer> {
+public interface CampañaFavoritaRepository extends JpaRepository<CampañaFavorita, Long> {
 }
