+++ src/main/java/pe/edu/upc/greenly/repositories/CampañaRepository.java
@@ -3,5 +3,5 @@ package pe.edu.upc.greenly.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pe.edu.upc.greenly.entities.Campaña;
 
-public interface CampañaRepository extends JpaRepository<Campaña,Integer> {
+public interface CampañaRepository extends JpaRepository<Campaña,Long> {
 }
