+++ src/main/java/pe/edu/upc/greenly/repositories/ComentarioRepository.java
@@ -3,5 +3,5 @@ package pe.edu.upc.greenly.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pe.edu.upc.greenly.entities.Comentario;
 
-public interface ComentarioRepository extends JpaRepository<Comentario, Integer> {
+public interface ComentarioRepository extends JpaRepository<Comentario, Long> {
 }
