+++ src/main/java/pe/edu/upc/greenly/repositories/DonanteRepository.java
@@ -3,5 +3,5 @@ package pe.edu.upc.greenly.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pe.edu.upc.greenly.entities.Donante;
 
-public interface DonanteRepository extends JpaRepository<Donante, Integer> {
+public interface DonanteRepository extends JpaRepository<Donante, Long> {
 }
