+++ src/main/java/pe/edu/upc/greenly/repositories/OngRepository.java
@@ -3,5 +3,5 @@ package pe.edu.upc.greenly.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pe.edu.upc.greenly.entities.Ong;
 
-public interface OngRepository extends JpaRepository<Ong, Integer> {
+public interface OngRepository extends JpaRepository<Ong, Long> {
 }
