+++ src/main/java/pe/edu/upc/greenly/service/ComentarioService.java
@@ -6,5 +6,5 @@ import java.util.List;
 public interface ComentarioService {
     public void insert(Comentario comentario);
     public List<Comentario> list();
-    public void delete(int id);
+    public void delete(Long id);
 }
